remove post and comment from main page componet and put it into common
