+++ frontend/src/MainPage/MainBox/User/Body.js
@@ -1,48 +1,14 @@
 import "./user.css";
-import CreatePost from "./CreatePost";
-import PostList from "./PostList";
-import React, { useState, useEffect } from 'react';
+import PostBox from "../../../Common/Post/PostBox";
 
-const Body = ({ id }) => {
-  const [body, setBody] = useState('');
+import React from 'react';
 
-  useEffect(() => {
-    const handleHashChange = () => {
-      const hash = window.location.hash.substring(1);
-      setBody(hash);
-    };
-
-    // Listen for hash changes in the URL
-    window.addEventListener('hashchange', handleHashChange);
-    handleHashChange(); // Initialize the body state based on the current hash
-
-    // Cleanup the event listener on unmount
-    return () => {
-      window.removeEventListener('hashchange', handleHashChange);
-    };
-  }, []);
-
-  const handleSubmitPost = (postData) => {
-    // Logic to handle the submission of the post data
-    setBody('postlist');
-    console.log('Submitted post:', postData);
-    // Make API requests or perform other operations here
-  };
+const Body = () => {
 
   return (
     <div className="main_body">
       {/* Conditional rendering based on the body state */}
-      {body === 'createpost' && (
-        <section id="createpost">
-          <CreatePost id={id} onSubmit={handleSubmitPost}/>
-        </section>
-      )}
-
-      {body === 'postlist' && (
-        <section id="postlist">
-          <PostList/>
-        </section>
-      )}
+      <PostBox/>
     </div>
   );
 };
